+++ src/main/java/org/eclipse/keyple/distributed/AbstractMessageHandlerAdapter.java
@@ -105,7 +105,18 @@ abstract class AbstractMessageHandlerAdapter {
    * @since 2.0.0
    */
   final void bindSyncNodeServer() {
-    node = new SyncNodeServerAdapter(this, 20);
+    bindSyncNodeServer(20);
+  }
+
+  /**
+   * Builds and bind a {@link SyncNodeServer} with the handler.<br>
+   * It must be invoked by the factory during the initialization phase.
+   *
+   * @param timeoutSeconds The timeout (in seconds).
+   * @since 2.4.0
+   */
+  final void bindSyncNodeServer(int timeoutSeconds) {
+    node = new SyncNodeServerAdapter(this, timeoutSeconds);
     isBoundToSyncNode = true;
   }
 
@@ -130,7 +141,19 @@ abstract class AbstractMessageHandlerAdapter {
    * @since 2.0.0
    */
   final void bindAsyncNodeServer(AsyncEndpointServerSpi endpoint) {
-    node = new AsyncNodeServerAdapter(this, endpoint, 20);
+    bindAsyncNodeServer(endpoint, 20);
+  }
+
+  /**
+   * Builds and bind a {@link AsyncNodeServer} with the handler.<br>
+   * It must be invoked by the factory during the initialization phase.
+   *
+   * @param endpoint The {@link AsyncEndpointServerSpi} endpoint.
+   * @param timeoutSeconds The timeout (in seconds).
+   * @since 2.4.0
+   */
+  final void bindAsyncNodeServer(AsyncEndpointServerSpi endpoint, int timeoutSeconds) {
+    node = new AsyncNodeServerAdapter(this, endpoint, timeoutSeconds);
     isBoundToSyncNode = false;
   }