+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Allows configuration of the timeout value used by server nodes (issue [#13]).
 ### Changed
 - Logging improvement.
 
@@ -45,6 +47,7 @@ It also brings many major API changes.
 [2.2.0]: https://github.com/eclipse-keyple/keyple-distributed-network-java-lib/compare/2.0.0...2.2.0
 [2.0.0]: https://github.com/eclipse-keyple/keyple-distributed-network-java-lib/releases/tag/2.0.0
 
+[#13]: https://github.com/eclipse-keyple/keyple-distributed-network-java-lib/issues/13
 [#6]: https://github.com/eclipse-keyple/keyple-distributed-network-java-lib/issues/6
 
 [eclipse-keyple/keyple#6]: https://github.com/eclipse-keyple/keyple/issues/6